+++ op_tests/triton_tests/attention/test_mha.py
@@ -9,7 +9,6 @@ from aiter.ops.triton.attention.mha import (
     flash_attn_varlen_func,
     mha_set_use_fused_bwd_kernel,
     mha_set_use_int64_strides,
-    mha_set_impl,
 )
 from aiter.ops.triton.attention.mha_v3 import (
     flash_attn_fp8_func,
@@ -1218,138 +1217,3 @@ def test_mha_backward_varlen_with_pe(
         rtol=bwd_rtol,
         msg=lambda msg: f"bwd dv mismatch\n\n{msg}\n",
     )
-
-
-@pytest.mark.parametrize(
-    "BATCH, SEQLEN_Q, SEQLEN_K, NUM_Q_HEADS, NUM_K_HEADS, HEAD_SZ, CAUSAL, VARLEN, BWD",
-    [
-        (1, 128, 128, 8, 8, 64, True, False, False),  # fwd causal
-        (2, 256, 256, 16, 4, 128, False, False, False),  # fwd GQA non-causal
-        (1, 128, 128, 8, 8, 64, True, True, False),  # fwd_varlen causal
-        (1, 128, 128, 8, 8, 64, True, False, True),  # bwd causal
-        (1, 128, 128, 8, 8, 64, True, True, True),  # bwd_varlen causal
-    ],
-)
-def test_mha_dao_ai(
-    BATCH: int,
-    SEQLEN_Q: int,
-    SEQLEN_K: int,
-    NUM_Q_HEADS: int,
-    NUM_K_HEADS: int,
-    HEAD_SZ: int,
-    CAUSAL: bool,
-    VARLEN: bool,
-    BWD: bool,
-    dtype=torch.float16,
-):
-    """Sanity test: verify dao_ai impl dispatch wiring for fwd/bwd x varlen."""
-    torch.cuda.empty_cache()
-    torch.manual_seed(20)
-    mha_set_impl("dao_ai")
-
-    q = torch.randn(
-        BATCH,
-        SEQLEN_Q,
-        NUM_Q_HEADS,
-        HEAD_SZ,
-        device="cuda",
-        dtype=dtype,
-        requires_grad=BWD,
-    )
-    k = torch.randn(
-        BATCH,
-        SEQLEN_K,
-        NUM_K_HEADS,
-        HEAD_SZ,
-        device="cuda",
-        dtype=dtype,
-        requires_grad=BWD,
-    )
-    v = torch.randn(
-        BATCH,
-        SEQLEN_K,
-        NUM_K_HEADS,
-        HEAD_SZ,
-        device="cuda",
-        dtype=dtype,
-        requires_grad=BWD,
-    )
-
-    if VARLEN:
-        query_padding_mask = generate_random_padding_mask(
-            SEQLEN_Q, BATCH, "cuda", mode="full"
-        )
-        key_padding_mask = generate_random_padding_mask(
-            SEQLEN_K, BATCH, "cuda", mode="full"
-        )
-        (
-            q_unpad,
-            k_unpad,
-            v_unpad,
-            cu_seqlens_q,
-            cu_seqlens_k,
-            max_seqlen_q,
-            max_seqlen_k,
-            q,
-            k,
-            v,
-            output_pad_fn,
-            dq_pad_fn,
-            dk_pad_fn,
-        ) = generate_qkv(q, k, v, query_padding_mask, key_padding_mask, kvpacked=False)
-        q_unpad.requires_grad_(BWD)
-        k_unpad.requires_grad_(BWD)
-        v_unpad.requires_grad_(BWD)
-
-        with torch.set_grad_enabled(BWD):
-            triton_out = flash_attn_varlen_func(
-                q_unpad,
-                k_unpad,
-                v_unpad,
-                cu_seqlens_q,
-                cu_seqlens_k,
-                max_seqlen_q,
-                max_seqlen_k,
-                causal=CAUSAL,
-            )
-    else:
-        query_padding_mask = None
-        key_padding_mask = None
-        with torch.set_grad_enabled(BWD):
-            triton_out = flash_attn_func(q, k, v, causal=CAUSAL)
-
-    # Forward check
-    torch_out, _, _ = attention_ref(
-        q,
-        k,
-        v,
-        causal=CAUSAL,
-        query_padding_mask=query_padding_mask,
-        key_padding_mask=key_padding_mask,
-    )
-    if VARLEN:
-        triton_out_padded = output_pad_fn(triton_out)
-        torch.testing.assert_close(triton_out_padded, torch_out, atol=1e-2, rtol=1e-2)
-    else:
-        torch.testing.assert_close(triton_out, torch_out, atol=1e-2, rtol=1e-2)
-
-    # Backward check
-    if BWD:
-        if VARLEN:
-            do = torch.randn_like(q)
-            triton_out = output_pad_fn(triton_out)
-            triton_dq, triton_dk, triton_dv = torch.autograd.grad(
-                triton_out, (q_unpad, k_unpad, v_unpad), do
-            )
-            triton_dq = dq_pad_fn(triton_dq)
-            triton_dk = dk_pad_fn(triton_dk)
-            triton_dv = dk_pad_fn(triton_dv)
-        else:
-            do = torch.randn_like(q)
-            triton_dq, triton_dk, triton_dv = torch.autograd.grad(
-                triton_out, (q, k, v), do
-            )
-        for name, grad in [("dq", triton_dq), ("dk", triton_dk), ("dv", triton_dv)]:
-            assert not grad.isnan().any(), f"dao_ai backward produced NaN in {name}"
-
-    mha_set_impl("default")
